Set up the SPI to be used by the ws2812_spi later

+++ microbit_v2/src/animations/carrousel.rs
@@ -50,7 +50,7 @@ impl Animation for Carrousel<'_> {
         }
 
         ws2812.write(self.data.borrow().iter().copied()).unwrap();
-        timer.delay_ms(settings.delay() as u32);
+        timer.delay_ms(settings.delay());
     }
 
     fn reset(&mut self) {
