+++ microbit_v2/src/animations/double_carrousel.rs
@@ -70,7 +70,7 @@ impl Animation for DoubleCarrousel<'_> {
         }
 
         ws2812.write(self.data.borrow().iter().copied()).unwrap();
-        timer.delay_ms(settings.delay() as u32);
+        timer.delay_ms(settings.delay());
     }
 
     fn reset(&mut self) {
