+++ microbit_v2/src/animations/uni_color_front_to_back_wave.rs
@@ -40,7 +40,7 @@ impl Animation for UniColorFrontToBackWave<'_> {
         self.position = (self.position + 1) % VERTICAL_SLICES.len();
 
         ws2812.write(self.data.borrow().iter().copied()).unwrap();
-        timer.delay_ms(settings.delay() as u32);
+        timer.delay_ms(settings.delay());
     }
 
     fn reset(&mut self) {
